+++ app/src/js/actions/index.js
@@ -424,6 +424,26 @@ export const bulkGranule = (payload) => ({
   }
 });
 
+export const bulkGranuleClearError = (requestId) => ({
+  type: types.BULK_GRANULE_CLEAR_ERROR,
+  requestId
+});
+
+export const bulkGranuleDelete = (payload) => ({
+  [CALL_API]: {
+    type: types.BULK_GRANULE_DELETE,
+    method: 'POST',
+    path: 'granules/bulkDelete',
+    requestId: payload.requestId,
+    body: payload.json
+  }
+});
+
+export const bulkGranuleDeleteClearError = (requestId) => ({
+  type: types.BULK_GRANULE_DELETE_CLEAR_ERROR,
+  requestId
+});
+
 export const deleteGranule = (granuleId) => ({
   [CALL_API]: {
     type: types.GRANULE_DELETE,