+++ CHANGELOG.md
@@ -18,10 +18,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **CUMULUS-1918**
   - Adds delete report button to the Reconciliation Reports page
 
+- **CUMULUS-1977**
+  - Added BulkGranuleModal component for creating modals to submit bulk granule requests
+
 ### Changed
 
 - **CUMULUS-1815**
   - Refactor some PDR components. No user facing changes.
+
+- **CUMULUS-1977**
+  - Updated BulkGranule component to display a modal that allows you to choose
+  the type of bulk request you want to submit: bulk granule operations or bulk
+  granule delete.
 
 - **CUMULUS-1994**
   - No default datetime filters are applied when the application is loaded.